Adds functions for making ZFS filesystems
This early draft work adds functions to morph to create a ZFS
root filesystem instead of the usual btrfs. It is not yet ready for
production but may be useful to see what is being worked on.

To use, you will need to have a Baserock system with ZFS and SPL
installed and running. You can then create a write extension for making
a raw disk image.